bugfix: use of reseting manually reference_target
the problem is due to of optimization, the reduced motorcycle according to the
polygon boarder is not reinstatiated but only modified according to the
intersection point with the polygon boarder. The problem with this approach was
that only the reference_point was resetted but the backup was not. This means
that the backup was set to the initial point which leads to that after using the
reset method on a MotorcycleSegment object the reference_target was set to the
initial endpoint, which introduced a bug that shuffle of motorcycles without a
intersection were not truncated at polygon boarder.